requirements.txt
streamlit
Pillow
torch
torchvision
transformers
sentence-transformers
easyocr
requests
python-dotenv